+++ src/pages/SignUpPage.tsx
@@ -1,17 +1,19 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import SignForm from "../components/common/form/SignForm";
-import useSignUp from "../hooks/auth/useSignUp";
+import useAuth from "../hooks/auth/useAuth";
 
 const SignUpPage = () => {
-  const { userInfo, handleSubmit, handleChange, isValid } = useSignUp();
+  const { userInfo, handleSubmit, handleChange, isValid } = useAuth();
+
+
   return (
     <div>
       <SignForm
         isValid={isValid}
         value={userInfo}
         title="회원가입"
-        handleSubmit={handleSubmit}
+        handleSubmit={(e) => handleSubmit(e, false)}
         handleChange={handleChange}
       />
       <Link to={"/signin"}>로그인 하러 가기</Link>
